requirements.txt
pyserial
pyyaml
RPi.GPIO
